Rename file to cssProp.
