+++ packages/app-tachyons-fela-prop/src/jsx-dev-runtime.js
@@ -1,5 +1,5 @@
 import { jsxDEV as _jsxDEV } from "react/jsx-dev-runtime";
-import { maybeExpandCssProp } from "./render.ts";
+import { maybeExpandCssProp } from "./cssProp.ts";
 
 export function jsxDEV(type, props, ...children) {
   return _jsxDEV(type, maybeExpandCssProp(props), ...children);
