+++ packages/app-tachyons-fela-prop/src/jsx-runtime.js
@@ -1,5 +1,5 @@
 import { jsx as _jsx } from "react/jsx-runtime";
-import { maybeExpandCssProp } from "./render.ts";
+import { maybeExpandCssProp } from "./cssProp.ts";
 
 export function jsx(type, props, ...children) {
   return _jsx(type, maybeExpandCssProp(props), ...children);
